refactor(lib/webidl2): modularize container types
