+++ lib/productions/helpers.js
@@ -2,6 +2,8 @@ import { Type } from "./type.js";
 import { Argument } from "./argument.js";
 import { Token } from "./token.js";
 import { ExtendedAttributes } from "./extended-attributes.js";
+import { Operation } from "./operation.js";
+import { Attribute } from "./attribute.js";
 
 /**
  * @param {string} identifier
@@ -150,4 +152,16 @@ export function return_type(tokeniser, typeName) {
     ret.type = "return-type";
     return ret;
   }
+}
+
+/**
+ * @param {import("../tokeniser").Tokeniser} tokeniser
+ */
+export function stringifier(tokeniser) {
+  const special = tokeniser.consume("stringifier");
+  if (!special) return;
+  const member = Attribute.parse(tokeniser, { special }) ||
+    Operation.parse(tokeniser, { special }) ||
+    tokeniser.error("Unterminated stringifier");
+  return member;
 }
