+++ test/invalid/baseline/id-underscored-number.txt
@@ -1,3 +1,3 @@
 Syntax error at line 1:
 interface _0 {};
-          ^ No name for interface
+          ^ Missing name in interface
