+++ test/invalid/baseline/inheritance-typeless.txt
@@ -1,0 +1,3 @@
+Syntax error at line 1, since `interface X`:
+interface X: {};
+             ^ Inheritance lacks a type
