+++ test/invalid/idl/inheritance-typeless.widl
@@ -1,0 +1,1 @@
+interface X: {};
